Avoids trying to stringify a huge array
When using `JSON.stringify` on a large array (200k items), it may
run out of memory and give an error. We run into this error at the
end of generation when we write to `build/json/_metadata.json`.

This commit makes it only stringify one item at a time, which is
probably slower, but avoids the error (could probably add logic to
only do this if needed).